refactor: extract mail and folder settings sections
